// apps/web/src/client/models/PostDetail.ts
export interface PostDetail {
    /**
     * The unique identifier of the post
     * @type {string}
     * @memberof PostDetail
     */
    id: string;
    /**
     * The user who posted the post
     * @type {SimpleUser}
     * @memberof PostDetail
     */
    user: SimpleUser;
    /**
     * The title of the post
     * @type {string}
     * @memberof PostDetail
     */
    title: string;
    /**
     * The tags of the post
     * @type {Array<string>}
     * @memberof PostDetail
     */
    tags: Array<string>;
    /**
     * The category of the post
     * @type {SimpleCategory}
     * @memberof PostDetail
     */
    category: SimpleCategory;
    /**
     * The answers of the post
     * @type {Array<AnswerWithQuestion>}
     * @memberof PostDetail
     */
    answers: Array<AnswerWithQuestion>;
    /**
     * The content of the post
     * @type {Array<SubPost>}
     * @memberof PostDetail
     */
    subPosts: Array<SubPost>;
    /**
     * The comments of the post
     * @type {Array<Comment>}
     * @memberof PostDetail
     */
    comments: Array<Comment>;
    /**
     * The feedback of the post
     * @type {Array<PostFeedback>}
     * @memberof PostDetail
     */
    feedback: Array<PostFeedback>;
}